+++ src/fields/core/fieldSubmit.vue
@@ -12,10 +12,13 @@ export default {
 	methods: {
 		onClick($event) {
 			if (this.schema.validateBeforeSubmit === true) {
+				// prevent a <form /> from having it's submit event triggered
+				// when we have to validate data first
+				$event.preventDefault();
 				let errors = this.$parent.validate();
 				let handleErrors = errors => {
 					if (!isEmpty(errors) && isFunction(this.schema.onValidationError)) {
-						this.schema.onValidationError(this.model, this.schema, errors);
+						this.schema.onValidationError(this.model, this.schema, errors, $event);
 					} else if (isFunction(this.schema.onSubmit)) {
 						this.schema.onSubmit(this.model, this.schema, $event);
 					}
@@ -27,6 +30,8 @@ export default {
 					handleErrors(errors);
 				}
 			} else if (isFunction(this.schema.onSubmit)) {
+				// if we aren't validating, just pass the onSubmit handler the $event
+				// so it can be handled there
 				this.schema.onSubmit(this.model, this.schema, $event);
 			}
 		}