+++ src/fields/optional/fieldSwitch.vue
@@ -123,7 +123,7 @@ $field-switch-height: 30px;
 	}
 	input:checked ~ .handle {
 		left: $field-switch-width - ($field-switch-height - 1px);
-		left: calc(100% - ($field-switch-height - 1px));
+		left: calc(100% - (#{$field-switch-height} - 1px));
 		box-shadow: -1px 1px 5px rgba(0, 0, 0, 0.2);
 	}